requirements.txt
sentence-transformers
qdrant-client
python-dotenv
tqdm
langchain-text-splitters
pandas
numpy
uuid
typing